translations: move TranslationForm to Views/
Components::TranslationForm -> Views::Controllers::Translations::Form
Per #4331. Two ERB callers (index.erb, edit.erb) updated.

Future opportunity (not in this PR): the form currently uses
FieldProxy with flat tag_<TAG> names to edit multiple TranslationString
records. Could be refactored to a FormObject envelope pattern (like
FormObject::ImageLicenseUpdates from #4323), but that's a controller-
shape change beyond a mechanical move.
