+++ app/views/controllers/translations/edit.erb
@@ -1,5 +1,5 @@
 <%= turbo_stream.update("translation_ui") do
-  concat(render(Components::TranslationForm.new(
+  concat(render(Views::Controllers::Translations::Form.new(
     lang: @lang, tag: @tag, edit_tags: @edit_tags,
     strings: @strings, for_page: @for_page,
     official_records: @official_records
